Separating component funcitons out into their respective files
